requirements.txt
psutil>=5.9.0
plyer>=2.1.0
pystray>=0.19.0
Pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
